feat(ops): migrate maintenance dispatcher, window, and watcher scan to v2
- Add maintenance_job_op.go: OperationDef "maintenance.job" runs any named
  maintenance job by ID; replaces queue.Enqueue in runMaintenanceJob handler
- Add scheduler_maintenance_window_op.go: OperationDef "maintenance.window"
  (12h timeout, ConcurrencyKey, IgnoreWindow param) runs all eligible tasks
- Update maintenance_dispatcher.go: EnqueueOp("maintenance.job") replacing
  queue.Enqueue; bump version 1.3.0 → 1.4.0
- Update scheduler_maintenance.go: EnqueueOp("maintenance.window") replacing
  queue.Enqueue; IgnoreWindow captured from ctx at call-site; bump 1.0.0 → 1.1.0
- Update server_lifecycle.go: file-watcher auto-scan uses opRegistry.EnqueueOp
  ("library.scan") directly (pure v2, no orphan v1 op); drop ulid import;
  bump 1.8.0 → 1.9.0
